Follow Pagination, Delays & Page Size Param
Thank you, @get-itips, for this! I stumbled across this, which proved extremely useful! I know very little PowerShell syntax, so excuse me if anything isn't as efficient as it could be!

## Bits I've added

- Saved the generated token as a global variable so that 2FA isn't required if the script is re-run
- While loop added to follow paginated pages
- Added delays to minimise API limits being hit
- Added some nicer error handling for when any API limits are hit
- Change the "pageSize" to be a passed argument
- Minor tweaks to remove repeated URLs and URIs being defined

An example of executing with the "pageSize" is: "TeamsClear 300" to use a page size of 300. I have set up a scripts directory on my machine, which I have added to my system path, as it allows me to simply call the cmd name and argument as stated above.